Module radeco::analysis::valueset

Structs

KnownBits
Value set described by bits set/cleared in the u64

SIntRange

UIntMultiple

UIntRange

Traits

ScannableSet
A value set over a totally ordered domain with methods to query the next greater or smaller value

ValueSet
A value set