Module radeco::frontend::cfg

This module is used to convert the instructions generated from esil into a CFG.

Control Flow Graphs (CFG) aid in the analysis and recovery of the program structure.

Structs

BasicBlock
A BasicBlock is the basic unit in a CFG. Every Instruction must be a part of one and only one BasicBlock. Every instruction in the BasicBlock executes sequentially. This BasicBlock implementation does not store the instructions themselves but rather just stores the indices of the instructions that belong to a block as a Range. Note that rust ranges are half-open. This means that the last value is not included as a part of this Range, i.e. (a..b) => [a, b).

CFG

Functions

make_graph