Struct radeco::frontend::cfg::BasicBlock

pub struct BasicBlock {
    pub range: Range<u64>,
    pub index: NodeIndex,
    pub reachable: bool,
    pub instructions: HashMap<String, Instruction>,
    pub label: String,
}

A BasicBlock is the basic unit in a CFG. Every Instruction must be a part of one and only one BasicBlock. Every instruction in the BasicBlock executes sequentially. This BasicBlock implementation does not store the instructions themselves but rather just stores the indices of the instructions that belong to a block as a Range. Note that rust ranges are half-open. This means that the last value is not included as a part of this Range, i.e. (a..b) => [a, b).

Fields

range

index

reachable

instructions

label