Module radeco::frontend::esil

Module to parse ESIL strings and convert them into the IR.

ESIL (Evaluable Strings Intermediate Language) is the IL used by radare2.

For a complete documentation of ESIL please check wiki.

Details

The Parser struct provides methods needed to convert a valid ESIL string into the IR. Parser::parse() parses the ESIL string and returns an Err if the ESIL string is Invalid.

Parser also provides Parser::emit_insts() to extract the Instructions it generates. Calling Parser::parse() several times will add more instructions.

Example

use radeco::frontend::esil;
let esil = "eax,ebx,^=";
let mut p = esil::Parser::new();
p.parse(esil);
for inst in &p.emit_insts() {
    println!("{}", inst);
}

Structs

Instruction

Operator

Parser

Value

Enums

Arity

Location

Opcode

ParseError